Improved RMQ performance, especially within the mediator
 * ship "raw" IP CCs for pcapdisk agency as IP packets with a
   pcap header, rather than encoding to ETSI just to have to
   decode them again to convert to pcap.
 * fix bad coding that meant we were declaring the rawip queue
   every time we wanted to publish to it.
 * pre-construct queue names in the collector recv threads for
   each known LIID, so we don't need to snprintf() every time we
   want to publish to a queue.
 * increase number of RMQ messages consumed by an LEA thread before
   sending an ack from 32 to 1024.
 * increase number of RMQ messages consumed by a collector
   recv thread before sending an ack from 1 to 32.
 * don't expire unwritten pcap messages in RMQ after 60 seconds,
   persist them as long as needed.
 * fix segfault when RMQ sends a message larger than the
   increment that we use to extend the receiving buffer in
   netcomms.c.
 * fix memory error when extracting the LIID from an exported
   message within a buffer larger than 64K.
 * send each encoded ETSI record as a separate message when
   using RMQ between the collector and mediator, rather than
   trying to batch multiple records in a single large message
   (rabbitmq is poor at dealing with larger messages).

+++ src/util.c
@@ -528,7 +528,7 @@ char *parse_iprange_string(char *ipr_str) {
  *  @return a pointer to the first character of the extracted LIID.
  */
 char *extract_liid_from_exported_msg(uint8_t *etsimsg,
-        uint16_t msglen, unsigned char *space, int maxspace,
+        uint64_t msglen, unsigned char *space, int maxspace,
         uint16_t *liidlen) {
 
     uint16_t l;
